split off wait_reapable into a separate module
